+++ src/room/Room.ts
@@ -434,8 +434,10 @@ class Room extends (EventEmitter as new () => TypedEmitter<RoomEventCallbacks>) 
         CriticalTimers.clearTimeout(connectTimeout);
         this.abortController?.signal.removeEventListener('abort', abortHandler);
         // also hook unload event
-        if (isWeb()) {
-          window.addEventListener('beforeunload', this.onBeforeUnload);
+        if (isWeb() && this.options.disconnectOnPageLeave) {
+          // capturing both 'pagehide' and 'beforeunload' to capture broadest set of browser behaviors
+          window.addEventListener('pagehide', this.onPageLeave);
+          window.addEventListener('beforeunload', this.onPageLeave);
           navigator.mediaDevices?.addEventListener('devicechange', this.handleDeviceChange);
         }
         this.setAndEmitConnectionState(ConnectionState.Connected);
@@ -592,7 +594,7 @@ class Room extends (EventEmitter as new () => TypedEmitter<RoomEventCallbacks>) 
     }
   }
 
-  private onBeforeUnload = async () => {
+  private onPageLeave = async () => {
     await this.disconnect();
   };
 
@@ -896,7 +898,8 @@ class Room extends (EventEmitter as new () => TypedEmitter<RoomEventCallbacks>) 
       this.audioContext = undefined;
     }
     if (isWeb()) {
-      window.removeEventListener('beforeunload', this.onBeforeUnload);
+      window.removeEventListener('beforeunload', this.onPageLeave);
+      window.removeEventListener('pagehide', this.onPageLeave);
       navigator.mediaDevices?.removeEventListener('devicechange', this.handleDeviceChange);
     }
     this.setAndEmitConnectionState(ConnectionState.Disconnected);
@@ -1318,8 +1321,14 @@ class Room extends (EventEmitter as new () => TypedEmitter<RoomEventCallbacks>) 
     this.emit(RoomEvent.TrackUnmuted, pub, this.localParticipant);
   };
 
-  private onLocalTrackPublished = (pub: LocalTrackPublication) => {
+  private onLocalTrackPublished = async (pub: LocalTrackPublication) => {
     this.emit(RoomEvent.LocalTrackPublished, pub, this.localParticipant);
+    if (pub.track instanceof LocalAudioTrack) {
+      const trackIsSilent = await pub.track.checkForSilence();
+      if (trackIsSilent) {
+        this.emit(RoomEvent.LocalAudioSilenceDetected, pub);
+      }
+    }
   };
 
   private onLocalTrackUnpublished = (pub: LocalTrackPublication) => {
@@ -1498,6 +1507,7 @@ export type RoomEventCallbacks = {
     publication: LocalTrackPublication,
     participant: LocalParticipant,
   ) => void;
+  localAudioSilenceDetected: (publication: LocalTrackPublication) => void;
   participantMetadataChanged: (
     metadata: string | undefined,
     participant: RemoteParticipant | LocalParticipant,